fix: correct featureImportance import after Saarthi rename
The module moved from lib/saarthi/ to lib/ but kept a two-level
relative import. Point it at ../api/inferenceContext and move the
leftover test next to the implementation.
